% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amalg.R
\name{amalg.ILR}
\alias{amalg.ILR}
\title{amalgamate two groups in a compositional data matrix to obtain ILR vector
Must input the log of the data matrix}
\usage{
amalg.ILR(Grp, LogData)
}
\arguments{
\item{Grp}{List containing two non-overlapping vectors of positive, non-zero integers less than the number of rows in the Data matrix.}

\item{LogData}{Logarithm of data matrix whose rows are parts and whose columns are samples.}
}
\description{
amalgamate two groups in a compositional data matrix to obtain ILR vector
Must input the log of the data matrix
}
