% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amalgamate.R
\name{amalgamate}
\alias{amalgamate}
\title{Amalgamation contrast function}
\usage{
amalgamate(Grps, TransformedData, pseudo.count = 0.65)
}
\arguments{
\item{Grps}{Either a list of \code{getPhyloGroups} or a single \code{getPhyloGroups} element}

\item{TransformedData}{Transformed Data. For amalgamation-phylofactorization, set \code{transform.fcn=I} and \code{contrast.fcn=amalgamate}.}

\item{pseudo.count}{Numeric greater than 0. If aggregation of a group is zero, it will be replaced with \code{pseudo.count}}
}
\description{
Amalgamation contrast function
}
\examples{
library(phylofactor)
set.seed(1)
m=50
n=20
tree <- rtree(m)
Grps <- getPhyloGroups(tree)
M <- matrix(rpois(m*n,1),nrow=m)
rownames(M) <- tree$tip.label
colnames(M) <- paste('Sample',1:n)

amalgamate(Grps,M)
amalgamate(Grps[[1]],M)
}
