% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomPlot.R
\name{atomPlot}
\alias{atomPlot}
\title{Color atoms on a plot of a phylogeny}
\usage{
atomPlot(atms, factor, tree = NULL, n = NULL, colorfcn = rainbow,
  type = "unrooted", show.tip.label = F, legend.return = F, ...)
}
\arguments{
\item{atms}{Either a list of atoms or a phylofactor object. See \code{\link{atoms}} or \code{\link{PhyloFactor}}}

\item{factor}{Integer. Atoms will be the set of atoms defined by factors up to the input \code{factor} in phylofactor object.}

\item{tree}{phylo-class tree. If input \code{atms} is a list of atoms, tree must be input, otherwise it will be extracted from phylofactor object.}

\item{n}{Size of total set to use for colorfcn - this enables multiple plots with differnent numbers of atoms to have the same color order.}

\item{colorfcn}{Color function. Default is rainbow.}

\item{type}{type of phylogeny for plot.phylo. Default is unrooted.}

\item{show.tip.label}{Logical. Whether or not to show tip labels. Default is False.}

\item{legend.return}{Logical indicating whether or not to return legend labels and colors for making a legend.}
}
\value{
plot.phylo where unique edges for each atoms are different colors, with optional return of legend information
}
\description{
Color atoms on a plot of a phylogeny
}
\examples{
library(phylofactor)
data('FTmicrobiome')

atomPlot(FTmicrobiome$PF,factor=3)
}

