% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomsphylofactor.R
\name{atoms.PhyloFactor}
\alias{atoms.PhyloFactor}
\title{Outputs the atoms made by the first n factors in PhyloFactor object}
\usage{
atoms.PhyloFactor(PF, nfactors = 1)
}
\description{
@param PF Phylofactor object. See \code{\link{PhylOFactor}}
 @param nfactors Number of factors, from 1:nfactors, whose resultant atoms are desired.
 @return a list of atoms
}

