% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceContrast.R
\name{BalanceContrast}
\alias{BalanceContrast}
\title{Project data matrix onto balance contrast (if data are log-data, then this will produce ILR balance)}
\usage{
BalanceContrast(Grp, TransformedData)
}
\arguments{
\item{Grp}{List containing two non-overlapping vectors of positive, non-zero integers less than the number of rows in the Data matrix.}

\item{TransformedData}{Transformed data matrix whose rows are parts and whose columns are samples.}
}
\description{
Project data matrix onto balance contrast (if data are log-data, then this will produce ILR balance)
}
