% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binPhyloPlot.R
\name{binPhyloPlot}
\alias{binPhyloPlot}
\title{Color bins on a plot of a phylogeny}
\usage{
binPhyloPlot(X, factor, tree = NULL, n = NULL, colorfcn = rainbow,
  type = "unrooted", show.tip.label = F, legend.return = F, ...)
}
\arguments{
\item{X}{Either a phylofactor object or a list of bins. See \code{\link{bins}} or \code{\link{PhyloFactor}}}

\item{factor}{Integer. Bins will be the set of bins defined by factors up to the input \code{factor} in phylofactor object.}

\item{tree}{phylo-class tree. If input \code{X} is a list of bins, tree must be input, otherwise it will be extracted from phylofactor object.}

\item{n}{Size of total set to use for colorfcn - this enables multiple plots with differnent numbers of bins to have the same color order.}

\item{colorfcn}{Color function. Default is rainbow.}

\item{type}{type of phylogeny for plot.phylo. Default is unrooted.}

\item{show.tip.label}{Logical. Whether or not to show tip labels. Default is False.}

\item{legend.return}{Logical indicating whether or not to return legend labels and colors for making a legend.}

\item{...}{optional input arguments to \code{\link{plot.phylo}}}
}
\value{
plot.phylo where unique edges for each bin are different colors, with optional return of legend information
}
\description{
Color bins on a plot of a phylogeny
}
\examples{
library(phylofactor)
data('FTmicrobiome')

binPhyloPlot(FTmicrobiome$PF,factor=3)
}
