% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binProjection.R
\name{binProjection}
\alias{binProjection}
\title{Projects phylofactored data onto bins defined by the factors 1:f}
\usage{
binProjection(PF, f, taxonomy = NULL, common.name = F, uniques = T,
  prediction = F)
}
\arguments{
\item{PF}{phylofactor object. See \link{PhyloFactor}}

\item{f}{integer between 0 and \code{PF$nfactors}}

\item{taxonomy}{Optional taxonomy. If input, the list of OTUs in each bin will be the taxonomic names, not the OTU IDs.}

\item{common.name}{Logical. If input taxonomy, will indicate whether to trim taxonomic list to the longest common prefix of each bin.}

\item{uniques}{Logical. If input taxonomy, will indicate whether to trim the taxonomic list to the unique names.}
}
\value{
Returns list containing the compositional dataset formed by the bins and the list of OTUs in each bin.
}
\description{
Projects phylofactored data onto bins defined by the factors 1:f
}

