% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsphylofactor.R
\name{bins.PhyloFactor}
\alias{bins.PhyloFactor}
\title{Outputs the bins made by the first n factors in PhyloFactor object
 @param PF Phylofactor object. See \code{\link{PhylOFactor}}
 @param nfactors Number of factors, from 1:nfactors, whose resultant bins are desired.
 @return a list of bins}
\usage{
bins.PhyloFactor(PF, nfactors = 1)
}
\description{
Outputs the bins made by the first n factors in PhyloFactor object
 @param PF Phylofactor object. See \code{\link{PhylOFactor}}
 @param nfactors Number of factors, from 1:nfactors, whose resultant bins are desired.
 @return a list of bins
}

