% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bins}
\alias{bins}
\title{Outputs bins from a partition matrix, V, or from a list of groups G and their full set.}
\usage{
bins(V = NULL, G = NULL, set = NULL)
}
\arguments{
\item{V}{Signed partition matrix, need not be normalized, but can be an ilr basis.}

\item{G}{If V is not input, can instead input a list of groups, G, whose elements are all contained in the input set}

\item{set}{If V is not input, set determines the full set of elements of which G is sub-groups and of which bins are constructed.}
}
\value{
a list of bins - list of elements corresponding to minimal, unsplit groups given the partition or the group list.
}
\description{
Outputs bins from a partition matrix, V, or from a list of groups G and their full set.
}
\examples{
V <- matrix(c(1,1,-1,-1,1,-1,0,0),byrow=F,ncol=2)
bins(V)

set=1:10
G <- list(c(1,2),c(1,2,3,4),c(8))
bins(G=G,set=set)
}
