% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTaxa.R
\name{binTaxa}
\alias{binTaxa}
\title{returns the taxonomic identity of taxa in bins}
\usage{
binTaxa(PF, nfactors, taxonomy, common.name = F, uniques = T)
}
\arguments{
\item{PF}{Phylofactor object}

\item{nfactors}{number of factors used (will look at bins created by factors \code{1:nfactors})}

\item{taxonomy}{Taxonomy: first column is OTU IDs and second column is taxonomy.}

\item{common.name}{whether to simplify bin-taxonomy to its longest-similar-prefix}

\item{uniques}{- whether to include all unique taxa}
}
\description{
returns the taxonomic identity of taxa in bins
}

