% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{internal \code{PhyloFactor} quality-control function}
\usage{
checkData(Data, tree, method)
}
\arguments{
\item{Data}{Data frame}

\item{tree}{phylo object}

\item{method}{See \code{\link{PhyloFactor}}}
}
\description{
internal \code{PhyloFactor} quality-control function
}
