% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorTaxa.R
\name{ColorTaxa}
\alias{ColorTaxa}
\title{Color tree based on taxonomic level}
\usage{
ColorTaxa(tree, Taxonomy, level = "p", legend = FALSE)
}
\arguments{
\item{tree}{Phylogeny whose tips are contained in the first column of the list, Taxonomy}

\item{Taxonomy}{list whose first column contains the tip labels of the tree and whose second column contains greengenes taxonomic strings for each OTU.}

\item{level}{Taxonomic level from set {'k','p','c','o','f','g','s'} default ='p'.}

\item{legend}{Whether ot not to include a legend of Taxonomic groups at that taxonomic level.}
}
\value{
colored tree labelled by taxa
}
\description{
Color tree based on taxonomic level
}

