% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decideMethod.R
\name{decideMethod}
\alias{decideMethod}
\title{decides Method automatically for phyca and twoSample data inputs}
\usage{
decideMethod(Data, X)
}
\arguments{
\item{Data}{\code{Data} input to \code{\link{PhyloFactor}}}

\item{X}{\code{X} input into \code{\link{PhyloFactor}}}
}
\description{
decides Method automatically for phyca and twoSample data inputs
}
