% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorFrame.R
\name{factorFrame}
\alias{factorFrame}
\title{Function for making factor-contrast in \code{\link{gpf}}}
\usage{
factorFrame(grp, spp)
}
\arguments{
\item{grp}{list containing two disjoint lists of species, such as thouse output from \code{\link{getGroups}}}

\item{spp}{set of species or tip-labels of tree.}
}
\description{
Function for making factor-contrast in \code{\link{gpf}}
}
