% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findunsplitgroups.R
\name{find.unsplit.Grps}
\alias{find.unsplit.Grps}
\title{Finds unsplit groups in a signed partition matrix, V}
\usage{
find.unsplit.Grps(V)
}
\arguments{
\item{V}{Partition matrix.}
}
\description{
Finds unsplit groups in a signed partition matrix, V
}
\examples{
V <- matrix(c(1,1,-1,-1,1,-1,0,0),byrow=FALSE,ncol=2)
find.unsplit.Grps(V)
}
