% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInterruptions.R
\name{findInterruptions}
\alias{findInterruptions}
\title{Finds interrupting edges along chain of edges}
\usage{
findInterruptions(edges, tree, original_community, new_community,
  old_edges = NULL, new_edges = NULL)
}
\arguments{
\item{edges}{non-branching chain of edges}

\item{tree}{phylo clas object}

\item{original_community}{original community}

\item{new_community}{new community - will look for edges from new community interrupting the tree from old community}

\item{old_edges}{optional input of old edges - the edges of master \code{tree} present in the old community}

\item{new_edges}{optional input of new edges - the edges of master \code{tree} present in new community.}
}
\description{
Finds interrupting edges along chain of edges
}

