% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWinner.R
\name{findWinner}
\alias{findWinner}
\title{Internal phylofactor function for finding the winning edge.}
\usage{
findWinner(nset, tree_map, treeList, treetips, Log.Data, choice, smallglm = F,
  frmla = NULL, X = NULL, ...)
}
\description{
Internal phylofactor function for finding the winning edge.
}

