% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM.R
\name{GAM}
\alias{GAM}
\title{internal generalized additive model method for \code{\link{PhyloFactor}} method='gam'}
\usage{
GAM(y, X, PF.output = FALSE, frmla, choice, ...)
}
\arguments{
\item{y}{ILR coordinate, y}

\item{X}{data frame, X}

\item{PF.output}{logical, if true will return customized output for phylofactor object}

\item{frmla}{formula for regression}
}
\description{
internal generalized additive model method for \code{\link{PhyloFactor}} method='gam'
}
