% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignedGroups.R
\name{getAssignedGroups}
\alias{getAssignedGroups}
\title{internal function for parallelized phylofactor}
\usage{
getAssignedGroups(nset, tree_map, treetips, treeList)
}
\arguments{
\item{nset}{set of nodes to be evaluated by worker}

\item{tree_map}{mapping of nodes to \code{treeList}}

\item{treetips}{number of tips in each tree}

\item{treeList}{list of phylo class objects.}
}
\description{
internal function for parallelized phylofactor
}
