% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveFunction.R
\name{getDeviance}
\alias{getDeviance}
\title{objective function for generalized phylofactorization}
\usage{
getDeviance(model, PartitioningVariables = "", ...)
}
\arguments{
\item{model}{model, such as glm, lm, nls, gam, which can be put into anova.}
}
\description{
objective function for generalized phylofactorization
}
