% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFactoredEdgesPAR.R
\name{getFactoredEdgesPAR}
\alias{getFactoredEdgesPAR}
\title{Parallelized version of getFactoredEdges}
\usage{
getFactoredEdgesPAR(cl = NULL, ncores = NULL, tree = NULL,
  V = NULL, PF = NULL)
}
\arguments{
\item{cl}{optional input \code{phyloFcluster}}

\item{ncores}{optional integer input ncores- must input wither cl or ncores.}

\item{tree}{phylo object}

\item{V}{partition matrix whose columns contain +/- 1 indicating group membership}

\item{PF}{phylofactor class object.}
}
\description{
Parallelized version of getFactoredEdges
}
\examples{
data("FTmicrobiome")
V <- FTmicrobiome$PF$basis
tree <- FTmicrobiome$PF$tree

cl <- phyloFcluster(2)

getFactoredEdgesPAR(cl,tree=tree,V=V[,1:10])
}
