% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupsPar.R
\name{getGroupsPar}
\alias{getGroupsPar}
\title{internal version of getGroups to allow memory-efficient parallelization}
\usage{
getGroupsPar(tree, n)
}
\arguments{
\item{tree}{phylo object}

\item{n}{node in \code{tree} whose descendants are used to get group/complement list.}
}
\description{
internal version of getGroups to allow memory-efficient parallelization
}
\examples{

library(ape)
set.seed(1)
tree <- rtree(15)
plot.phylo(tree,show.tip.label=F,use.edge.length=F)
tiplabels()
nodelabels()
}

