% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabelledGrp.R
\name{getLabelledGrp}
\alias{getLabelledGrp}
\title{Internal phylofactor function - get labelled Group from phylofactor}
\usage{
getLabelledGrp(factor, tree, Groups)
}
\arguments{
\item{factor}{integer indicating the element of \code{Groups} to be labelled}

\item{tree}{phylo object - global phylogeny used to indicate whether the elements of \code{Groups[factor]} are monophyletic}

\item{Groups}{list of group/complement pairs corresponding to the split bin}
}
\description{
Internal phylofactor function - get labelled Group from phylofactor
}
\examples{
library(ape)
set.seed(6)
tree <- rtree(10)
Groups <- getGroups(tree)
treeList <- list(tree)
binList <- list(1:10)
factor <- 16
grp <- getLabelledGrp(factor,tree,Groups)
grp
}

