% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewGroups.R
\name{getNewGroups}
\alias{getNewGroups}
\title{Internal phylofactor function - get new grouplist corresponding to atomlist & treelist}
\usage{
getNewGroups(tree, treeList, atomList)
}
\arguments{
\item{tree}{phylo object - global tree}

\item{treeList}{list of subtrees for each atom in \code{atomList}}

\item{atomList}{list of atoms corresponding to current factorization}
}
\value{
List of groups for amalgamation. For example, see \code{updateAtomList}
}
\description{
Internal phylofactor function - get new grouplist corresponding to atomlist & treelist
}

