% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignal.R
\name{getSignal}
\alias{getSignal}
\title{Replicate objective from phylofactor object}
\usage{
getSignal(grp, PF)
}
\arguments{
\item{grp}{phylo-group list}

\item{PF}{phylofactor object}
}
\description{
Replicate objective from phylofactor object
}
