% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Grabs P value and F statistic from aov of glm object.}
\usage{
getStats(GLM, y = NULL)
}
\arguments{
\item{GLM}{glm object}

\item{y}{dependent variable - used when inputting bigglm object.}
}
\value{
vector containing P value and F statistic from aov(GLM)
}
\description{
Grabs P value and F statistic from aov of glm object.
}
\examples{
x <- rnorm(10)
y <- 2*x+rnorm(10)

GLM <- glm(y~x)
getStats(GLM)
}

