% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMean.R
\name{gMean}
\alias{gMean}
\title{Computes geometric mean using logarithms}
\usage{
gMean(Y, MARGIN = 2)
}
\arguments{
\item{Y}{matrix or vector}

\item{MARGIN}{input for \code{\link{apply}}}
}
\description{
Computes geometric mean using logarithms
}
