% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGTplot.R
\name{HGTplot}
\alias{HGTplot}
\title{Visualize results form simHGT}
\usage{
HGTplot(sim, main = "Trait Evolution", cex = 1.5, lwd = 2, pch = 16,
  ...)
}
\arguments{
\item{sim}{Simulation output from \code{\link{simHGT}}}

\item{main}{title of plot}

\item{cex}{cex applied across plot, points and lines.}

\item{lwd}{line width}

\item{pch}{point choice, see \code{\link{plot}}}
}
\description{
Visualize results form simHGT
}
\examples{
library(phylofactor)

set.seed(5)
tree <- rtree(5)
sim <- simHGT(tree,0.3)
par(mfrow=c(2,1))
plot.phylo(tree)
HGTplot(sim,lwd=2,main='HGT Dynamics')
}
