% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilrInv.R
\name{ilrInv}
\alias{ilrInv}
\title{Function to take inverse of ilr transform}
\usage{
ilrInv(x, V)
}
\arguments{
\item{x}{balances}

\item{V}{balancing elements}
}
\description{
Function to take inverse of ilr transform
}
