% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilrvec.R
\name{ilrvec}
\alias{ilrvec}
\title{Internal function for computing an ilr balancing element given a list of split taxa in and the total number of taxa}
\usage{
ilrvec(ix, n)
}
\arguments{
\item{ix}{two-member list containing the parts split by a partition}

\item{n}{total number of parts}
}
\value{
ILR basis element corresponding to the split of ix[[1]] with ix[[2]] out of n-part composition.
}
\description{
Internal function for computing an ilr balancing element given a list of split taxa in and the total number of taxa
}
\examples{
n <- 10
ix <- NULL
ix[[1]] <- c(3,4)
ix[[2]] <- c(5,6,7,8)
v1 <- ilrvec(ix,n)
sum(v1)

ix[[1]] <- c(5,6)
ix[[2]] <- c(7,8)
v2 <- ilrvec(ix,n)
sum(v2)
v1 \%*\% v2
}
