% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTaxa.R
\name{listTaxa}
\alias{listTaxa}
\title{Lists all unique taxonomic names in a Taxonomy at a given level}
\usage{
listTaxa(s, minimum.level = 1, common.name = F, uniques = F)
}
\arguments{
\item{s}{Character vector of taxonomic names. First row is species IDs, second row is their taxonomy. This function assumes greengenes-compatible taxonomic strings}

\item{minimum.level}{Integer. Minimum number of taxonomic levels to include counting from highest to lowest.}

\item{common.name}{logical whether or not to return the common name in the list of taxa}

\item{uniques}{logical whether or not to trim taxonomic names to only unique entries}
}
\value{
list of taxonomic strings for all taxa up to speficied level.
}
\description{
Lists all unique taxonomic names in a Taxonomy at a given level
}
\examples{
data(FTmicrobiome)
sapply(FTmicrobiome$taxonomy[,2],as.character) \%>\%
   listTaxa(minimum.level=2,uniques=TRUE)
}
