% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mAggregation.R
\name{mAggregation}
\alias{mAggregation}
\title{marginally stable aggregation of binomial data}
\usage{
mAggregation(Data, grp, tree, MetaData, expfamily, frmla)
}
\arguments{
\item{Data}{matrix or list of matrices containing LHS variables to be aggregated}

\item{grp}{see output of \code{\link{getPhyloGroups}}}

\item{tree}{phylo object}

\item{MetaData}{meta-data table containing "sample" and variables found in formula}

\item{expfamily}{Character, currently either "binomial" or "gaussian", indicating family of exponential family for marginally-stable aggregation.}

\item{frmla}{formula - the response variable (or cbind(Successes,Failures) for \code{expfamily='binomial'},) will be aggregated.}
}
\description{
marginally stable aggregation of binomial data
}
