% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcat.R
\name{matcat}
\alias{matcat}
\title{Function for aligning rows and concatenating two matrices.}
\usage{
matcat(A, B)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix}
}
\value{
Matrix where upper-left block is matrix A, upper-right block is the abundances in B of parts in A. Lower left block is abundances in A of parts in B not found in A, and lower-right block is abundances in B of parts not found in A.
}
\description{
Function for aligning rows and concatenating two matrices.
}
\examples{
A <- matrix(rnorm(12),ncol=4)
rownames(A) <- 1:3
B <- matrix(rnorm(36),ncol=6)
rownames(B) <- sample(6)
matcat(A,B)
}

