% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.to.phyloframe.R
\name{matrix.to.phyloframe}
\alias{matrix.to.phyloframe}
\title{Converts matrix to phyloframe}
\usage{
matrix.to.phyloframe(Mat, X = NULL, data.name = "Data", empty.val = NA)
}
\arguments{
\item{Mat}{Matrix. Rownames must be species and colnames must be samples found in \code{X$Sample}}

\item{X}{data.table or data.frame of meta-data to include in phyloframe}

\item{data.name}{Character. Name of data from \code{Mat} for phyloframe}

\item{empty.val}{Values of \code{Mat} to ignore and not include in phyloframe}
}
\description{
Converts matrix to phyloframe
}
\examples{
library(phylofactor)
set.seed(1)
m=3   #number of species
n=5   #number of samples
d=10  #number of data points
DF <- data.frame('Species'=sample(letters[1:m],d,replace=TRUE),
                 'Sample'=sample(1:n,d,replace=TRUE),
                 'Abundance'=rpois(d,4))
Mat <- phyloframe.to.matrix(DF,mat.data='Abundance',empty.val=NA)
DF2 <- matrix.to.phyloframe(Mat,data.name='Abundance')

X <- data.frame('Sample'=as.character(1:n),'x'=2*(1:n),'y'=-(1:n))
matrix.to.phyloframe(Mat,X,data.name='Abundance')
}
