% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTUtoTaxa.R
\name{OTUtoTaxa}
\alias{OTUtoTaxa}
\title{Input OTUs and outputs their taxonomic detail, optinally trimmed to their common taxonomic levels.}
\usage{
OTUtoTaxa(otus, Taxonomy, common.name = F, uniques = F,
  minimum.level = Inf)
}
\arguments{
\item{otus}{Species IDs found in first column of Taxonomy}

\item{Taxonomy}{Taxonomy whose first column has OTUs/species IDs and whose second column has greengenes-compatible taxonomic strings}

\item{common.name}{Logical indicating whether or not to trim the output to the longest common prefix in the taxonomic strings of all otus}

\item{uniques}{Logical whether or not to trim list to only unique entries}

\item{minimum.level}{integer indicating the minimum number of taxonomic levels to include starting from the coarsest (Left-most) taxonomic level.}
}
\description{
Input OTUs and outputs their taxonomic detail, optinally trimmed to their common taxonomic levels.
}
