% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.BINprojection.R
\name{pf.BINprojection}
\alias{pf.BINprojection}
\title{Projects phylofactored data onto bins defined by the factors 1:factor}
\usage{
pf.BINprojection(PF, factor = PF$nfactors, taxonomy = NULL,
  common.name = F, uniques = T, prediction = F, rel.abund = F)
}
\arguments{
\item{PF}{phylofactor object. See \link{PhyloFactor}}

\item{factor}{integer between 0 and \code{PF$nfactors}}

\item{taxonomy}{Optional taxonomy. If input, the list of OTUs in each bin will be the taxonomic names, not the OTU IDs.}

\item{common.name}{Logical. If input taxonomy, will indicate whether to trim taxonomic list to the longest common prefix of each bin.}

\item{uniques}{Logical. If input taxonomy, will indicate whether to trim the taxonomic list to the unique names.}

\item{prediction}{Logical. If true, will output the predictions for each bin instead of the observations.}

\item{rel.abund}{Logical. If true, will return the total relative abundances in the bin. Default, F, returns the geometric mean of the relative abundances of each bin}
}
\value{
Returns list containing the compositional dataset formed by the bins and the list of OTUs in each bin.
}
\description{
Projects phylofactored data onto bins defined by the factors 1:factor
}
