% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.crossValidateMap.R
\name{pf.crossValidateMap}
\alias{pf.crossValidateMap}
\title{Function for mapping phylogenetic factors in a phylofactor object to a new community through universal tree}
\usage{
pf.crossValidateMap(pf, universal.tree, new.community,
  factors = 1:pf$nfactors)
}
\arguments{
\item{pf}{phylofactor class object whose tree tip-labels are all in the \code{universal.tree}}

\item{new.community}{New community}

\item{factors}{which factors to cross-map}

\item{univeral.tree}{Universal phylogeny containing all the tip-labels of \code{pf} and all elements of \code{new.community}}
}
\description{
Function for mapping phylogenetic factors in a phylofactor object to a new community through universal tree
}
