% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.flow.R
\name{pf.flow}
\alias{pf.flow}
\title{Generates flow of phylofactor object}
\usage{
pf.flow(PF, factor = PF$nfactors)
}
\arguments{
\item{PF}{phylofactor class object}

\item{factor}{integer to which factor the flow should detail}
}
\description{
Generates flow of phylofactor object
}
