% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.getGroups.R
\name{pf.getGroups}
\alias{pf.getGroups}
\title{Gets Groups considered at a given factor level}
\usage{
pf.getGroups(PF, factor)
}
\arguments{
\item{PF}{phylofactor class object from \code{\link{PhyloFactor}}}

\item{factor}{factor level of interest, must be greater than 1. For factor=1, use \code{\link{getGroups}}. A list of paired groups considered at that level will be returned.}
}
\value{
output similar to \code{\link{getGroups}}
}
\description{
Gets Groups considered at a given factor level
}
