% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.groupsTospecies.R
\name{pf.groupsTospecies}
\alias{pf.groupsTospecies}
\title{Converts integer-indexed groups to species-labelled groups in phylofactor object}
\usage{
pf.groupsTospecies(pf)
}
\arguments{
\item{pf}{phylofactor class object from \code{\link{PhyloFactor}}, \code{\link{twoSampleFactor}} or \code{\link{gpf}}}
}
\description{
Converts integer-indexed groups to species-labelled groups in phylofactor object
}
\examples{
library(phylofactor)
data(FTmicrobiome)
pf <- FTmicrobiome$PF
species.groups <- pf.groupsTospecies(pf)
pf$factors[1:3,]
#the otuIDs/species in factor 3, Group1 are:
species.groups[[3]][[1]]
}
