% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.heatmap.R
\name{pf.heatmap}
\alias{pf.heatmap}
\title{Heatmap for phylofactor data}
\usage{
pf.heatmap(PF = NULL, tree = NULL, Data = NULL, factors = NULL,
  column.order = NULL, branch.length = "none", color = NA,
  font.size = 0, ...)
}
\arguments{
\item{PF}{phylofactor class object from either \code{\link{PhyloFactor}} or \code{\link{gpf}} with \code{algorithm='mStable'}}

\item{tree}{optional input, only used if \code{PF} is not input. If inputting \code{tree}, must also input \code{Data}}

\item{Data}{Dataset with rows equal to the number of species. Ideally, rownames are all equal to \code{PF$tree} of \code{tree} tip-labels. Otherwise, rows must be in order of tree tip-labels.}

\item{factors}{vector of integer factors for input with \code{PF}. Will use \code{\link{pf.tree}} to highlight clades obtained at each factor.}

\item{column.order}{vector of integers of length equal to \code{ncol(Data)} or \code{ncol(PF$Data)}. Will re-order the data (e.g. according to desired meta-data).}

\item{branch.length}{input for \code{\link{ggtree}}}

\item{color}{input \code{color} for \code{gheatmap}}

\item{font.size}{font size for \code{gheatmap}}

\item{...}{additional arguments passed to \code{\link{gheatmap}}}
}
\description{
Heatmap for phylofactor data
}
\examples{
library(phylofactor)
require(ggpubr)
data(FTmicrobiome)
PF <- FTmicrobiome$PF
obs <- PF$transform.fcn(PF$Data)
pred <- predict(PF,factors=1:3)
observed <- pf.heatmap(PF,Data=obs,factors=1:3,width=3)
predicted <- pf.heatmap(PF,Data=pred,factors=1:3,width=3)
ggarrange(observed,predicted,nrow=2)
}
