% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.predict.R
\name{pf.predict}
\alias{pf.predict}
\title{Predicts based on phylofactorization}
\usage{
pf.predict(PF, factors = NULL, ...)
}
\arguments{
\item{PF}{PhyloFactor object}

\item{factors}{Set of factors to be used for prediction. Currently, must be a continuous integer sequence from 1:nfactors}

\item{...}{optional input arguments for glm}
}
\value{
estimated data based on estimates from phylofactorization
}
\description{
Predicts based on phylofactorization
}
\examples{
library(compositions)
library(caper)
set.seed(1)
tree <- unroot(rtree(8))
Data <- matrix(rlnorm(80),nrow=8)
rownames(Data) <- tree$tip.label
Grp <- clade.members.list(tree)$'12'
for (gg in Grp){
Data[gg,] <- Data[gg,]*20^(seq(-1,1,length.out=10))
}

T <- 1:10
colnames(Data) <- T

Data <- t(rcomp(t(Data)))
rownames(Data) <- tree$tip.label
phylo.heatmapAW(tree,clr(t(Data)))

PF <- PhyloFactor(Data,tree,X=T,nclades=1)

CommunityEst <- pf.predict(PF,factors=1,newdata=data.frame(X=seq(0,11,by=1/3)))
rownames(CommunityEst) <- rownames(Data)

phylo.heatmapAW(tree,clr(t(CommunityEst)))
}
