% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.resume.R
\name{pf.resume}
\alias{pf.resume}
\title{Resume phylofactorization for a phylofactor object}
\usage{
pf.resume(pf, ...)
}
\arguments{
\item{pf}{phylofactor object from \code{\link{PhyloFactor}}, \code{\link{PhyCA}}, \code{\link{twoSamplefactor}} or \code{\link{gpf}}}

\item{...}{additional arguments..... INCOMPLETE}
}
\description{
Resume phylofactorization for a phylofactor object
}
