% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.taxa.R
\name{pf.taxa}
\alias{pf.taxa}
\title{summarizes taxonomic composition at given factor}
\usage{
pf.taxa(pf, taxonomy, factor = 1)
}
\arguments{
\item{pf}{phylofactor object}

\item{taxonomy}{taxonomy whose first column is tree tip labels in phylofactor object and whose second column is semicolon-delimited taxonomy}

\item{factor}{positive integer. Which factor to summarize.}
}
\description{
summarizes taxonomic composition at given factor
}
\examples{
library(phylofactor)
data(FTmicrobiome)
pf <- FTmicrobiome$PF
tx <- FTmicrobiome$taxonomy

pf.taxa(pf,tx,1)
}
