% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.tidy.R
\name{pf.tidy}
\alias{pf.tidy}
\title{Concise summarization of phylofactor summary}
\usage{
pf.tidy(smry)
}
\arguments{
\item{smry}{PF summary class object. see \code{\link{phylofactor.summary}}}
}
\value{
list containing concise taxa IDs of groups split, as well as their predicted ratios of abundances in different conditions
}
\description{
Concise summarization of phylofactor summary
}
\examples{
library(phylofactor)
data('FTmicrobiome')

smry <- phylofactor.summary(FTmicrobiome$PF,FTmicrobiome$taxonomy,factor=1)

pf.tidy(smry)
}
