% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyCA.R
\name{PhyCA}
\alias{PhyCA}
\title{Performs Phylogenetic Principal Components Analysis}
\usage{
PhyCA(Data, tree, ncores = NULL, ncomponents = NULL, output.edges = T,
  tol = 1e-05, quiet = T)
}
\arguments{
\item{Data}{Positive-valued data matrix whose rownames are tip-labels in the input \code{tree}.}

\item{tree}{phylo-class object whose tip-labels cover the rownames of Data.}

\item{ncores}{optional number of cores for built-in parallelization. Be cautious of memory - each worker is sent a copy of the dataset and tree, producing ncores+1 copies of the dataset & tree.}

\item{output.edges}{Logical, whether or not to output the edges in the input \code{tree} corresponding to phylogenetic components}

\item{tol}{tolerance for compositional matrix. Rounding error in large datasets can lead to columns of compositional data not summing to 1.}

\item{quiet}{Logical, whether or not to quiet warnings.}
}
\value{
PhyCA object containing Data, tree, basis, and edges of the phylogeny corresponding to each split.
}
\description{
Performs Phylogenetic Principal Components Analysis
}
\examples{

library(phylofactor)
data("FTmicrobiome")

Data <- FTmicrobiome$PF$Data
tree <- FTmicrobiome$PF$tree
X <- FTmicrobiome$X
taxonomy <- FTmicrobiome$taxonomy
phytools::phylo.heatmap(tree,t(compositions::clr(t(Data))))


phca <- PhyCA(Data,tree,ncomponents = 2)
phcaPAR <- PhyCA(Data,tree,ncomponents=3,ncores=3)

phylofactor.visualize(phcaPAR,X=X,dimension=3)

phytools::phylo.heatmap(tree,t(compositions::clr(t(Data))))
edgelabels(text=1:3,edge = unlist(phcaPAR$edges),cex=2)
}

