% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phca.tidy.R
\name{phyca.tidy}
\alias{phyca.tidy}
\title{clean description of phca object}
\usage{
phyca.tidy(phca, Taxonomy, ncomponents = NULL, taxa.split = F,
  common.name = F, uniques = T, getEdges = F, ncores = NULL, ...)
}
\arguments{
\item{phca}{\code{PhyCA} object}

\item{Taxonomy}{Greengenes taxonomy, first column are otus and second column are taxonomic strings of format "p__Phylum; c__Class; o__Order..."}

\item{ncomponents}{Integer number of PhyCA components to tidy, from \code{1:ncomponents}}

\item{taxa.split}{Logical whether to summarize taxonomic detail of component by only including the taxa names which are unique to each group.}

\item{common.name}{Logical, whether or not to trim output taxonomic summary to the longest common prefix. Default \code{FALSE}}

\item{uniques}{Logical, whether or not to trim taxonomic output to unique taxonomic names. Default \code{TRUE}}

\item{getEdges}{Logical, whether or not to get edges corresponding to the splits. Default \code{FALSE}}

\item{ncores}{Integer number of cores to use if getting edges - does not affect runtime if \code{getEdges=FALSE}}

\item{...}{optional input arguments to \code{\link{OTUtoTaxa}}}
}
\description{
clean description of phca object
}
\examples{

library(phylofactor)
data("FTmicrobiome")

Data <- FTmicrobiome$PF$Data
tree <- FTmicrobiome$PF$tree
taxonomy <- FTmicrobiome$taxonomy

phca <- PhyCA(Data,tree,ncomponents = 3,ncores=2,output.edges=FALSE)

#the standard
td <- phyca.tidy(phca,taxonomy)

#or, for a simpler view
td <- phyca.tidy(phca,taxonomy,taxa.split=TRUE)

sum(FTmicrobiome$PF$basis[,1:3]-phca$basis[,1:3])  
#the first three phylofactors here correspond to the first three ILR-phylogenetic Components. 
}
