% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.frame.to.matrix.R
\name{phylo.frame.to.matrix}
\alias{phylo.frame.to.matrix}
\title{converts data frame in appropriate phylofactor format to matrix.}
\usage{
phylo.frame.to.matrix(DF)
}
\arguments{
\item{DF}{data frame or data table containing columns "Sample", "Species", and "N". "N" is the observed data for generalized phylofactorization.}
}
\description{
converts data frame in appropriate phylofactor format to matrix.
}
\examples{
library(phylofactor)
DF <- data.frame('Sample'=rep(1:3,each=3),'Species'=rep(c('a','b','c'),times=3),'N'=1:9)
phylo.frame.to.matrix(DF)
}
