% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloheatmapAW.R
\name{phylo.heatmapAW}
\alias{phylo.heatmapAW}
\title{Ghetto, AW version of phylo.heatmap}
\usage{
phylo.heatmapAW(tree, Y, tipLabels = TRUE, ...)
}
\arguments{
\item{tree}{phylogeny (ape class)}

\item{Y}{data matrix. row names must match tip labels of tree}

\item{tipLabels}{Logical indicating whether or not to include tip labels}

\item{...}{optional arguments for plot.phylo(...)}
}
\description{
Ghetto, AW version of phylo.heatmap
}
\examples{
set.seed(1)
tree <- unroot(rtree(7))
X <- as.factor(c(rep(0,5),rep(1,5)))
sigClades <- Descendants(tree,c(9,12),type='tips')
Data <- matrix(rlnorm(70,meanlog = 8,sdlog = .5),nrow=7)
rownames(Data) <- tree$tip.label
colnames(Data) <- X
Data[sigClades[[1]],X==0] <- Data[sigClades[[1]],X==0]*8
Data[sigClades[[2]],X==1] <- Data[sigClades[[2]],X==1]*9
Data <- t(clo(t(Data)))

es1 <- extractEdges(tree,tree$tip.label[sigClades[[1]]],type=3)
es2 <- extractEdges(tree,tree$tip.label[sigClades[[2]]],type=3)
ecs <- rep('black',Nedge(tree))
ecs[es1] <- 'red'
ecs[es2] <- 'blue'
phylo.heatmapAW(tree,Data,tipLabels=F,edge.width=4,edge.color=ecs)
nodelabels(c(9,12),c(9,12),cex=2)
}

