% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylofactor.projection.R
\name{phylofactor.projection}
\alias{phylofactor.projection}
\title{Projects Data onto basis defined by phylofactor object}
\usage{
phylofactor.projection(PF, Data = NULL, nfactors = 2)
}
\arguments{
\item{PF}{Phylofactor object. See \code{\link{PhyloFactor}}}

\item{Data}{Optional new data matrix, Must have rownames matching those in PF$Data}

\item{nfactors}{Number of factors used for projection. Default nfactors=2 for ordination-visualization}
}
\description{
Projects Data onto basis defined by phylofactor object
}

