% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylofactorTaxaIDs.R
\name{phylofactor.TaxaIDs}
\alias{phylofactor.TaxaIDs}
\title{Returns the taxonomic IDs for all the OTUs in the atoms formed by splits in a Phylofactor object.}
\usage{
phylofactor.TaxaIDs(PF, Taxonomy, tree, clades = NULL, common.name = F)
}
\arguments{
\item{PF}{Phylofactor object}

\item{Taxonomy}{taxonomy list. First column must be species IDs corresponding to tip-labels of the tree, second column must be taxonomy strings (currently, this works for greengenes taxonomy strings - beware other taxonomy formats)}

\item{tree}{Phylogeny}

\item{clades}{Clades, or factors, of phylofactor object to be inlcuded. Currently, this must be a sequence from 1:n, i.e. cannot be c(2,3,4...,n) or have any missing clades. The reason for this is the atom function currently requires input of a sequential partition.}

\item{common.name}{Logical indicating whether or not to truncate the output to only one taxonomic assignment per atom, based on the finest, common name to all taxa in the atom.}
}
\description{
Returns the taxonomic IDs for all the OTUs in the atoms formed by splits in a Phylofactor object.
}

