% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylofactor.visualize.R
\name{phylofactor.visualize}
\alias{phylofactor.visualize}
\title{Phylofactor Ordination-Visualization}
\usage{
phylofactor.visualize(PF, X = NULL, dimension = 2, default.colors = T,
  Legend = T, lx = NULL, ly = NULL, colorbar.name = NULL,
  colorbar.ticks = 5, ...)
}
\arguments{
\item{PF}{Phylofactor object}

\item{X}{independent variable for colormap - must be the same length as dim(PF$Data)[2]}

\item{dimension}{Either 2 or 3 - the dimension of the ordination-visualization plot}

\item{default.colors}{Logical indicating if samples should be colored using default settings. If default.colors=F, user must provide appropriate color arguments for plot or scatterplot3d, and no legend will be made.}

\item{Legend}{Logical indicating if legend should be included. Must also have default.colors=T}

\item{lx}{x location for legend. If \code{dimension} = 2, x must be a numeric. If \code{dimension} = 3, x must be an scatterplot3d()$xyz.convert list}

\item{ly}{y location for legend, only applicable if \code{dimension}=2}

\item{colorbar.name}{The name of the colorbar for quantitative independent variable, \code{X}. If independent variable is not a factor, \code{phylofactor.visualization} assumes they are quantitative and will make a colorbar on the sorted independent variable.}

\item{colorbar.ticks}{Number of ticks in colorbar}

\item{...}{Additional arguments. If \code{dimension==2}, these are arguments for \code{\link{plot}}. Otherwise, they are additional arguments for \code{\link{scatterplot3d}}}
}
\description{
Phylofactor Ordination-Visualization
}
\examples{
data(FTmicrobiome)
phylofactor.visualize(FTmicrobiome$PF)
phylofactor.visualize(FTmicrobiome$PF,dimension=3)

pf <- FTmicrobiome$PF
X <- 1:40
phylofactor.visualize(pf,dimension=3,colorbar.name='sample number')
}

