% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloFcluster.R
\name{phyloFcluster}
\alias{phyloFcluster}
\title{Produces cluster object with necessary libraries to perform phylofactorization}
\usage{
phyloFcluster(ncores = 2, ...)
}
\arguments{
\item{ncores}{number of cores}

\item{...}{optional input arguments for \code{\link{makeCluster}}}
}
\description{
Produces cluster object with necessary libraries to perform phylofactorization
}
