% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloFrame.R
\name{phyloFrame}
\alias{phyloFrame}
\title{Tacks "phylo" variable onto data frame for \code{\link{gpf}}}
\usage{
phyloFrame(Data, grp, tree)
}
\arguments{
\item{Data}{data table input to \code{\link{gpf}}. Must have column \code{"Species"}}

\item{grp}{Group list element from \code{\link{getPhyloGroups}}}

\item{tree}{phylo class object}
}
\description{
Tacks "phylo" variable onto data frame for \code{\link{gpf}}
}
\examples{
library(phylofactor)
set.seed(1)
tree <- rtree(10)
Data <- data.table('Species'=tree$tip.label,'x'=rnorm(10),key='Species')
Grps <- getPhyloGroups(tree)
phylofactor:::phyloFrame(Data,Grps[[1]],tree)

grp <- list(c(1:3),c(5:6))
phyloFrame(Data,grp,tree)
}
