% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloframe.to.matrix.R
\name{phyloframe.to.matrix}
\alias{phyloframe.to.matrix}
\title{Converts input data frame for \link{gpf} to matrix for faster indexing}
\usage{
phyloframe.to.matrix(DF, mat.data = "N", empty.val = 0)
}
\arguments{
\item{DF}{data frame containing, at the minimum, columns \code{"Species"}, \code{"Sample"}, and the column with character string defined in \code{mat.data}. If \code{DF} does not contain \code{"Sample"}, a vector will be output.}

\item{mat.data}{Character string. Column of \code{DF} input to the matrix}

\item{empty.val}{Value for empty Species x Sample pairs. Default is 0.}
}
\description{
Converts input data frame for \link{gpf} to matrix for faster indexing
}
\examples{
library(phylofactor)
m=10   #number of species
n=15   #number of samples
d=100  #number of data points
DF <- data.frame('Species'=sample(letters[1:m],d,replace=TRUE),
                 'Sample'=sample(1:n,d,replace=TRUE),
                 'Abundance'=rpois(d,2))
M <- phyloframe.to.matrix(DF,mat.data='Abundance',empty.val=0)

}
