% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloreg.R
\name{phyloreg}
\alias{phyloreg}
\title{Returns regression of x on y according to frmla}
\usage{
phyloreg(y, x, frmla, ...)
}
\arguments{
\item{y}{response variable}

\item{x}{independent variable}

\item{frmla}{Formula for dependence of y on x}
}
\value{
glm object
}
\description{
Returns regression of x on y according to frmla
}
\examples{
x <- rnorm(10)
y <- x+rnorm(10)
phyloreg(y,x,y~x)
}

