% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloRegression.R
\name{PhyloRegression}
\alias{PhyloRegression}
\title{PhyloFactor internal function to prep, find and summarize the best ILR coordinate}
\usage{
PhyloRegression(TransformedData, X, frmla, Grps = NULL, contrast.fcn = NULL,
  choice, treeList = NULL, cl, totalvar = NULL, ix_cl, treetips = NULL,
  grpsizes = NULL, tree_map = NULL, quiet = T, nms = NULL,
  smallglm = F, choice.fcn, method = "glm", ...)
}
\arguments{
\item{TransformedData}{Transformed data matrix whose rows are species and columns are samples.}

\item{X}{independent variables for input into glm.}

\item{frmla}{Formula for input into glm by lapply(Y,FUN = pglm,x=xx,frmla=frmla,choice,...).}

\item{Grps}{Groups - a list whose elements are two-element lists containing the groups and their compliments for amalgamation into ILR coordinates}

\item{contrast.fcn}{Function taking as input \code{contrast.fcn(Grps,TransformedData)} and returning a matrix with rows corresponding to \code{Grps} and columns corresponding to columns in \code{TransformedData}. Must work for both list of \code{Grps} and a single element, e.g. \code{Grps[[1]]}. For advanced phylofactorization with customized \code{choice.fcn}, the output from \code{contrast.fcn} is input directly into \code{choice.fcn}.}

\item{choice}{Choice function for determining the group maximizing the objective function. Currently the only allowable inputs are 'var' - minimize residual varaince - and 'F' - minimize test-statistic from anova.}

\item{treeList}{List of trees formed by phylofactorization and cutting trees along factored edges.}

\item{cl}{\code{phyloFcluster} for built-in parallelization of phylofactorization calculations.}

\item{totalvar}{Total variance of dataset.}

\item{ix_cl}{Cluster split of nodes in treeList}

\item{treetips}{Number of tips in treeList for quickly identifying whether nodes correspond to a root}

\item{grpsizes}{Number of nodes in each tree of treeList}

\item{tree_map}{Cumulative number of nodes in trees of treeList - allows rapid mapping of nodes in ix_cl to appropriate tree in treeList.}

\item{quiet}{Logical to supress warnings}

\item{nms}{rownames of TransformedData, allowing reliable mapping of rows of data to tree.}

\item{smallglm}{Logical. See \code{\link{PhyloFactor}}}

\item{choice.fcn}{optional customized choice function to choose 'best' edge; see \code{\link{PhyloFactor}}}

\item{method}{See \code{\link{PhyloFactor}}}

\item{...}{optional input arguments for \code{\link{glm}}}
}
\description{
PhyloFactor internal function to prep, find and summarize the best ILR coordinate
}
