% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsigclades.R
\name{plot.sig.clades}
\alias{plot.sig.clades}
\title{Plots tree and highlights otus given by sigOTUs}
\usage{
\method{plot}{sig.clades}(tree, sigOTUs, Taxonomy, level = "p",
  legend = FALSE, ...)
}
\arguments{
\item{tree}{phylogeny (ape class)}

\item{sigOTUs}{taxa to highlight, corresponding to tip labels of tree.}

\item{Taxonomy}{Taxonomy whose first column is OTU ids and whose second column is a greengenes style taxonomic string}

\item{level}{Taxonomic level from {'k','p','c','o','f','g'} for coloring tree}

\item{legend}{Logical whether or not to include legend of taxonomic types corresponding to colors in tree}

\item{...}{optional arguments for plot.phylo()}
}
\value{
colored plot of phylogeny, highlighting sigOTUs
}
\description{
Plots tree and highlights otus given by sigOTUs
}
\examples{
data("FTmicrobiome")
set.seed(1)
tree <- drop.tip(tree,tree$tip.label[sample(Ntip(tree),Ntip(tree)-100)])
otus <- tree$tip.label
Taxonomy <- FTmicrobiome$taxonomy[which(FTmicrobiome$taxonomy[,1] \%in\% otus),]
txa <- listTaxa(Taxonomy,level='c')
txa <- txa[[3]]  #these are the p__Firmicutes; c__Bacilli
otus <- TaxaToOTU(txa,Taxonomy)[[1]]

plot.sig.clades(tree,otus,Taxonomy,level='p',legend=T)
}

