% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictAmalgam.R
\name{PredictAmalgam}
\alias{PredictAmalgam}
\title{Internal function for prediction of data matrix given a method of amalgamation and its expectation}
\usage{
PredictAmalgam(yhat, Grp, n)
}
\arguments{
\item{yhat}{Estimated coordinate (expected value of coordinate) from regression on the amalgamation of Grp}

\item{Grp}{Two element list containing a group and its complement for amalgamation & contrast}

\item{n}{number of parts in entire dataset}

\item{method}{Method for amalgamation.}

\item{Pbasis}{coming soon}
}
\description{
Internal function for prediction of data matrix given a method of amalgamation and its expectation
}

