% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.phylofactor.R
\name{print.phylofactor}
\alias{print.phylofactor}
\title{Prints phylofactor class objects}
\usage{
print.phylofactor(PF)
}
\arguments{
\item{PF}{phylofactor object}
}
\description{
Prints phylofactor class objects
}
