% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.phylofactor.summary.R
\name{print.phylofactor.summary}
\alias{print.phylofactor.summary}
\title{Prints phylofactor.summary objects}
\usage{
\method{print}{phylofactor.summary}(s)
}
\arguments{
\item{output}{from \code{\link{summary.phylofactor}}}
}
\description{
Prints phylofactor.summary objects
}
