% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvDeviance.R
\name{pvDeviance}
\alias{pvDeviance}
\title{obtains deviance for partitioning variables - default objective function for \code{\link{gpf}}}
\usage{
pvDeviance(model, grp, tree, PartitioningVariables = "")
}
\arguments{
\item{model}{model, such as glm, lm, nls, gam, which can be put into anova.}

\item{grp}{two-member list of integers, see \code{\link{getPhyloGroups}}.}

\item{tree}{phylo class object}

\item{PartitioningVariables}{string - used to find partitioning variables in anova table.}
}
\description{
obtains deviance for partitioning variables - default objective function for \code{\link{gpf}}
}
