% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.R
\name{rarefy}
\alias{rarefy}
\title{Rarefies and OTU table to a given level}
\usage{
rarefy(OTUTable, to = "min", method = "multinom")
}
\arguments{
\item{OTUTable}{Table of species abundances/counts whose rows are species and whose columns are samples}

\item{to}{Number of counts post-rarefaction, from 'min','max', or an integer. Default is 'min', minimum number of counts of all samples in the table}

\item{method}{Method for randomly drawing counts. Currently only supports "multinom", but coming soon: 'nbinom' and 'PoisLN'}
}
\value{
Table of abundances with colSums(Table) = to number specified.
}
\description{
Rarefies and OTU table to a given level
}

