% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeGroup.R
\name{removeGroup}
\alias{removeGroup}
\title{Slow, brute-force AF internal function for removing a group from a list of groups/complements}
\usage{
removeGroup(Grps, id)
}
\arguments{
\item{Grps}{Full set of groups and their complements, in format of output of \code{\link{getGroups}}}

\item{id}{Group number to be removed from Grps}
}
\value{
list of groups and their complements constrained to the partition defined by removing Grps[[id]]
}
\description{
Slow, brute-force AF internal function for removing a group from a list of groups/complements
}
\examples{
set.seed(1)
tree <- rtree(10)
Grps <- getGroups(tree)
removeGroup(Grps,11)
}

