% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simHGT.R
\name{simHGT}
\alias{simHGT}
\title{Simulates brownian motion with HGT between random clades via poisson pt process}
\usage{
simHGT(tree, rate, a = 0, mu = 0, sig2 = 1)
}
\arguments{
\item{tree}{phylo class object}

\item{rate}{positive real number indicating rate of horizontal gene transfer per unit edge-length in tree}

\item{a}{real number ancestral state}

\item{mu}{drift}

\item{sig2}{volatility}
}
\description{
Simulates brownian motion with HGT between random clades via poisson pt process
}
\examples{
library(phylofactor)
library(ape)

tree <- rtree(10)
sim <- simHGT(tree,0.01,2,-3,0.5)
}
