% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.group.R
\name{summary.group}
\alias{summary.group}
\title{Summarize groups from phylofactor}
\usage{
\method{summary}{group}(PF, tree, taxonomy, factor, grp)
}
\arguments{
\item{PF}{Phylofactor output}

\item{tree}{The tree}

\item{taxonomy}{The taxonomy.}

\item{factor}{The factor.}

\item{grp}{The group.}
}
\description{
This is the description
}

