% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.phylofactor.r
\name{summary.phylofactor}
\alias{summary.phylofactor}
\title{Summarize phylofactor output}
\usage{
\method{summary}{phylofactor}(PF, tree, taxonomy, node = NULL,
  factor = NULL, subtree = F, prediction = T, tipLabels = F, ...)
}
\description{
Summarize phylofactor output
}

