% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaSplit.R
\name{TaxaSplit}
\alias{TaxaSplit}
\title{Describes the taxa split at a particular factor summarized by summary.phylofactor}
\usage{
TaxaSplit(Summary)
}
\arguments{
\item{Summary}{summary.phylofactor object}
}
\value{
list of taxa (with trimmed taxonomy strings) unique to the group & complement for the input summarized factor
}
\description{
Describes the taxa split at a particular factor summarized by summary.phylofactor
}

