% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaToOTU.R
\name{TaxaToOTU}
\alias{TaxaToOTU}
\title{Quick conversino from taxonomic string to OTU ID}
\usage{
TaxaToOTU(Taxa, taxonomy)
}
\arguments{
\item{Taxa}{Taxonomic string, e.g. k__Animalia or 'Animal'}

\item{taxonomy}{Greengenes format taxonomy whose first column is OTUids and whose second column is taxonomic strings}
}
\value{
list of OTU IDs whose taxonomic string contains "Taxa" input.
}
\description{
Quick conversino from taxonomic string to OTU ID
}

