% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeILR.R
\name{tree.ilr}
\alias{tree.ilr}
\title{Computes ilr basis corresponding to a phylogeny.}
\usage{
tree.ilr(phylo)
}
\arguments{
\item{phylo}{Ape class phylo object. Must be ultrametric - if not, we recommend trying phylo <- chronopl(phylo,lambda=1) before inputting.}
}
\description{
Computes ilr basis corresponding to a phylogeny.
}
\examples{
library(ape)
library(compositions)
tree <- rtree(10)
V <- tree.ilr

Data <- matrix(rpois(100,lambda=20),nrow=5)
Data[Data==0] <- 0.65

# Get the ILR coordinates for a dataset
X <- t(ilr(t(Data),V))

}

