% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueTaxa.R
\name{uniqueTaxa}
\alias{uniqueTaxa}
\title{Returns the first unique taxonomic level, starting from K, for all otus in x compared to y}
\usage{
uniqueTaxa(x, y)
}
\arguments{
\item{x}{list of taxonomic strings whose unique taxonomies we want to know}

\item{y}{list of taxonomic strings for "subtraction" from x}
}
\value{
list of taxonomic strings unique to x, i.e. the shortest taxonomic prefix of x that is unique to x
}
\description{
Returns the first unique taxonomic level, starting from K, for all otus in x compared to y
}

