% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueTips.R
\name{uniqueTips}
\alias{uniqueTips}
\title{Find all unique tips in a set of clades}
\usage{
uniqueTips(tree, Clades)
}
\arguments{
\item{tree}{ape class phylogeny}

\item{Clades}{clades of interest}
}
\value{
a vector of all the tips that are unique to \code{Clades} in \code{tree}
}
\description{
Find all unique tips in a set of clades
}
\examples{
set.seed(1)
tree <- rtree(10)
plot.phylo(tree)
nodelabels()
Clades <- c(16,13)

tips <- uniqueTips(tree,Clades)
tiplabels('***',tips,cex=3)
}

