% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAtomList.R
\name{updateAtomList}
\alias{updateAtomList}
\title{Updates Atom list given a split group}
\usage{
updateAtomList(atomList, grp)
}
\arguments{
\item{atomList}{List of atoms}

\item{grp}{Group indicating the split of an atom}
}
\description{
Updates Atom list given a split group
}
\examples{
library(ape)
set.seed(6)
tree <- rtree(10)
Groups <- getGroups(tree)
treeList <- list(tree)
atomList <- list(1:10)
factor <- 16
grp <- getLabelledGrp(factor,tree,Groups)
grp

treeList <- updateTreeList(treeList,atomList,grp,tree)
atomList <- updateAtomList(atomList,grp)


Groups <- getNewGroups(tree,treeList,atomList)

factor=2
grp <- getLabelledGrp(factor,tree,Groups)
grp
unlist(grp)
tree$tip.label[unlist(grp)]

treeList <- updateTreeList(treeList,atomList,grp,tree)
atomList <- updateAtomList(atomList,grp)



Groups <- getNewGroups(tree,treeList,atomList)
factor = 11
grp <- getLabelledGrp(factor,tree,Groups)
grp
unlist(grp)
tree$tip.label[unlist(grp)]

treeList <- updateTreeList(treeList,atomList,grp,tree)
atomList <- updateAtomList(atomList,grp)
}

