% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBinList.R
\name{updateBinList}
\alias{updateBinList}
\title{Updates Bin list given a split group}
\usage{
updateBinList(binList, grp)
}
\arguments{
\item{binList}{List of bins}

\item{grp}{Group indicating the split of an bin}
}
\description{
Updates Bin list given a split group
}
\examples{
library(ape)
set.seed(6)
tree <- rtree(10)
Groups <- getGroups(tree)
treeList <- list(tree)
binList <- list(1:10)
factor <- 16
grp <- getLabelledGrp(factor,tree,Groups)
grp

treeList <- updateTreeList(treeList,binList,grp,tree)
binList <- updateBinList(binList,grp)


Groups <- getNewGroups(tree,treeList,binList)

factor=2
grp <- getLabelledGrp(factor,tree,Groups)
grp
unlist(grp)
tree$tip.label[unlist(grp)]

treeList <- updateTreeList(treeList,binList,grp,tree)
binList <- updateBinList(binList,grp)



Groups <- getNewGroups(tree,treeList,binList)
factor = 11
grp <- getLabelledGrp(factor,tree,Groups)
grp
unlist(grp)
tree$tip.label[unlist(grp)]

treeList <- updateTreeList(treeList,binList,grp,tree)
binList <- updateBinList(binList,grp)
}
