% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VAR}
\alias{VAR}
\title{Computes variance for \code{\link{PhyloFactor}} method='max.var'}
\usage{
VAR(y, X, PF.output = FALSE, ...)
}
\arguments{
\item{y}{ILR coordinate, y}

\item{X}{data frame, X}

\item{PF.output}{logical, if true will return customized output for phylofactor object}

\item{frmla}{formula for regression}
}
\description{
Computes variance for \code{\link{PhyloFactor}} method='max.var'
}
