% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichAtomSplit.R
\name{whichAtomSplit}
\alias{whichAtomSplit}
\title{Internal function - id which atom in atomList split}
\usage{
whichAtomSplit(grp, atomList)
}
\arguments{
\item{grp}{split atom}

\item{atomList}{list of atoms}
}
\value{
index where atomList[[ix]] contains grp
}
\description{
Internal function - id which atom in atomList split
}

