% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichBinSplit.R
\name{whichBinSplit}
\alias{whichBinSplit}
\title{Internal function - id which bin in binList split}
\usage{
whichBinSplit(grp, binList)
}
\arguments{
\item{grp}{split bin}

\item{binList}{list of bins}
}
\value{
index where binList[[ix]] contains grp
}
\description{
Internal function - id which bin in binList split
}
