% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amalgamate.R
\name{amalgamate}
\alias{amalgamate}
\title{Amalgamation contrast function}
\usage{
amalgamate(Grps, TransformedData)
}
\arguments{
\item{Grps}{Either a list of \code{getPhyloGroups} or a single \code{getPhyloGroups} element}

\item{TransformedData}{Transformed Data. For amalgamation-phylofactorization, set \code{transform.fcn=I} and \code{contrast.fcn=amalgamate}.}
}
\description{
Amalgamation contrast function
}
\examples{
library(phylofactor)
set.seed(1)
m=10
n=5
tree <- rtree(m)
Grps <- getPhyloGroups(tree)
M <- matrix(rlnorm(m*n),nrow=m)
rownames(M) <- tree$tip.label
colnames(M) <- paste('Sample',1:n)

amalgamate(Grps,M)
amalgamate(Grps[[1]],M)
}
