% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amalgamate.R
\name{amalgamate}
\alias{amalgamate}
\title{amalgamate two groups in a compositional data matrix}
\usage{
amalgamate(Grp, Data, method, collapse = F)
}
\arguments{
\item{Grp}{List containing two non-overlapping vectors of positive, non-zero integers less than the number of rows in the Data matrix.}

\item{Data}{Compositional data matrix whose rows are parts and whose columns are samples.}

\item{method}{Method for amalgamation. Takes values of "sum" and "ILR". ILR amalgamates using geometric means, whereas "sum" amalgamates by summing relative abundances.}

\item{collapse}{Logical indicating whether or not to collapse groups (collapse=T) by summing or geometric means, or to output log-ratios from amalgamation & comparison.}
}
\value{
If collapse=T, returns either a log-ratio of the groups. If collapse=F, returns the geometric mean of each group.
}
\description{
amalgamate two groups in a compositional data matrix
}
\examples{
library(compositions)
set.seed(1)
Data <- matrix(rlnorm(80),nrow=8)
Grp <- list(c(1,2),c(3,4,8))
for (gg in Grp[[1]]){
Data[gg,] <- Data[gg,]*10^(seq(-1,1,length.out=10))
}

Data <- t(rcomp(t(Data)))

method='ILR'
lr <- amalgamate(Grp,Data,method)
plot(1:10,lr,xlab='sample',ylab='log-ratio of Grp {1,2} over {3,4,8}')
}
