% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorTaxa.R
\name{ColorTaxa}
\alias{ColorTaxa}
\title{Colors tree edges based on taxanomic ID}
\usage{
ColorTaxa(tree, taxonomy, level = "p", outputlegend = F, legend = FALSE,
  ...)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{taxonomy}{list whose first column is OTUs and second column is green-genes style taxonomy, e.g. k__Bacteria; p__Bacteroidetes}

\item{level}{String indicating taxonomic level for coloring. Must be in set {k,p,c,o,f,g,s}}

\item{...}{additional arguments for plot.phylo}
}
\description{
Colors tree edges based on taxanomic ID
}
\examples{
data(FTmicrobiome)
tree <- FTmicrobiome$PF$tree
taxonomy <- FTmicrobiome$taxonomy

ColorTaxa(tree,taxonomy,level='p',legend=T,show.tip.label=FALSE,type='unrooted')



L <- ColorTaxa(tree,taxonomy,level='c',legend=F,show.tip.label=FALSE,type='unrooted',outputlegend=T)

lims <- par('usr')
legend(lims[1],lims[4],legend=L$Taxa,fill=L$colors,cex=.6)
}

