% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorTaxa.R
\name{ColorTaxa}
\alias{ColorTaxa}
\title{Colors tree edges based on taxanomic ID}
\usage{
ColorTaxa(tree, taxonomy, level = "p", outputlegend = F, colorfcn = NULL,
  legend = FALSE, scramble = F, ...)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{taxonomy}{list whose first column is OTUs and second column is green-genes style taxonomy, e.g. k__Bacteria; p__Bacteroidetes}

\item{level}{String indicating taxonomic level for coloring. Must be in set {k,p,c,o,f,g,s}}

\item{outputlegend}{Logical indicating whether or not to output the list of taxa and their corresponding colors for manual generation of a legend}

\item{colorfcn}{Color Palette for taxa. Default is rainbow.}

\item{legend}{Logical whether or not to include legend. Most trees produce pretty ugly legends due to large numbers of taxa, so default is \code{FALSE}}

\item{scramble}{Logical - whether or not to scramble the order of taxa-color associations by a random draw. Useful for re-drawing colors in case closely related taxa are difficult to distinguish because of adjacent colors in colorfcn.}

\item{...}{additional arguments for plot.phylo}
}
\description{
Colors tree edges based on taxanomic ID
}
\examples{
data(FTmicrobiome)
tree <- FTmicrobiome$PF$tree
taxonomy <- FTmicrobiome$taxonomy

ColorTaxa(tree,taxonomy,level='p',legend=T,show.tip.label=FALSE,type='unrooted')



L <- ColorTaxa(tree,taxonomy,level='c',legend=F,show.tip.label=FALSE,type='unrooted',outputlegend=T)

lims <- par('usr')
legend(lims[1],lims[4],legend=L$Taxa,fill=L$colors,cex=.6)
}

