% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWinner.R
\name{findWinner}
\alias{findWinner}
\title{Internal PhyloRegression function for finding the winning edge.}
\usage{
findWinner(nset, tree_map, treeList, treetips, contrast.fcn = NULL, choice,
  method = "glm", smallglm = F, frmla = NULL, xx = NULL,
  choice.fcn = NULL, ...)
}
\arguments{
\item{nset}{set of nodes}

\item{tree_map}{mapping cumulative number of nodes in treeList, used to map elements of nset to their appropriate tree in treeList.}

\item{treeList}{list containing disjoint trees from phylofactor / PhyCA}

\item{treetips}{number of tips in each tree}

\item{contrast.fcn}{See \code{\link{PhyloFactor}} or example functions \code{\link{BalanceContrast}}, \code{\link{amalgamate}}}

\item{choice}{string indicating how we choose the winner. Must be either \code{'var'}, \code{'F'}, or \code{'phyca'}}

\item{method}{See \code{\link{PhyloFactor}}}

\item{smallglm}{Logical - whether or not to use regular \code{glm}. if smallglm=F, will use \code{\link{bigglm}} from the \code{\link{biglm}} package.}

\item{frmla}{Formula for \code{\link{glm}}. See \code{\link{PhyloFactor}} for more details.}

\item{xx}{data frame containing non-ILR (\code{Data}) variables used in \code{frmla}}

\item{choice.fcn}{See \code{\link{PhyloFactor}}}

\item{...}{optional input arguments to \code{\link{glm}}}
}
\description{
Internal PhyloRegression function for finding the winning edge.
}
