% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewGroups.R
\name{getNewGroups}
\alias{getNewGroups}
\title{Internal phylofactor function - get new grouplist corresponding to binlist & treelist}
\usage{
getNewGroups(tree, treeList, binList)
}
\arguments{
\item{tree}{phylo object - global tree}

\item{treeList}{list of subtrees for each bin in \code{binList}}

\item{binList}{list of bins corresponding to current factorization}
}
\value{
List of groups for amalgamation. For example, see \code{updateBinList}
}
\description{
Internal phylofactor function - get new grouplist corresponding to binlist & treelist
}

