% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjective.R
\name{getObjective}
\alias{getObjective}
\title{objective function for \code{\link{gpf}}}
\usage{
getObjective(grp, tree, DF, X, SampleFrame, size = 1, frmla)
}
\arguments{
\item{grp}{list containing two disjoint lists of species, such as thouse output from \code{\link{getGroups}}}

\item{tree}{phylo class object}

\item{DF}{data table containing Species, observations (N), and sample}

\item{X}{meta-data containing variables in formula}
}
\description{
objective function for \code{\link{gpf}}
}
