% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjective.R
\name{getObjective}
\alias{getObjective}
\title{objective function for \code{\link{gpf}}}
\usage{
getObjective(grp, tree, Data, X, binom.size, frmla, expfamily = "gaussian",
  model.fcn = stats::glm, PartitioningVariables = "", ...)
}
\arguments{
\item{grp}{list containing two disjoint lists of species, such as thouse output from \code{\link{getGroups}}}

\item{tree}{phylo class object}

\item{Data}{matrix whose rows are species and columns are samples}

\item{X}{meta-data containing variables in formula}
}
\description{
objective function for \code{\link{gpf}}
}
