% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjective.R
\name{getObjective}
\alias{getObjective}
\title{objective function for \code{\link{gpf}}}
\usage{
getObjective(grp, tree, Data, frmla, MetaData = NULL,
  PartitioningVariables = "", mStableAgg, expfamily = "gaussian",
  model.fcn = stats::glm, objective.fcn = pvDeviance,
  ignore.tips = F, ...)
}
\arguments{
\item{grp}{list containing two disjoint lists of species, such as thouse output from \code{\link{getGroups}}}

\item{tree}{phylo class object}

\item{Data}{If \code{mStableAgg==TRUE}, a matrix whose rows are species and columns are samples. Otherwise, a data table whose columns include "Species" and "Sample" and whose key is "Species".}

\item{frmla}{formula for \code{model.fcn}}

\item{MetaData}{meta-data containing variables in formula and the column "Sample". If \code{mStableAgg==F}, this input is not used - all variables must be contained in \code{Data}}

\item{PartitioningVariables}{character vector containing of interest for phylofactorization partitioning.}

\item{mStableAgg}{logical. See \code{\link{gpf}}}

\item{expfamily}{character string indicating manner of m-stable aggregation for \code{\link{mAggregation}}. Only "binomial" is meaningfully different.}

\item{model.fcn}{model function, such as \code{\link{glm}} or \code{gam}.}

\item{objective.fcn}{Objective function taking output from \code{model.fcn} as input. See \code{\link{gpf}}.}

\item{...}{additional arguments for \code{model.fcn}}
}
\description{
objective function for \code{\link{gpf}}
}
