% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpf.R
\name{gpf}
\alias{gpf}
\title{Generalized phylofactorization - currently skeleton for binomial glm}
\usage{
gpf(DF, X, tree, frmla, nfactors, ncores = NULL, size = 1)
}
\arguments{
\item{DF}{data table containing columns of "Species", 'N' (counts, 0<=N<=3), "sample".}

\item{X}{meta-data containing "sample" and variables found in \code{frmla}}

\item{nfactors}{integer for number of factors to find}

\item{ncores}{integers for number of cores to use for parallelization}

\item{size}{size of binomial samples for each species. Size=1 for presence/absence data.}
}
\description{
Generalized phylofactorization - currently skeleton for binomial glm
}
