% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpf.R
\name{gpf}
\alias{gpf}
\title{Generalized phylofactorization - currently skeleton for binomial glm; will expand to exponential, gamma etc. soon.}
\usage{
gpf(Data, tree, X, frmla, nfactors = NULL, ncores = NULL, binom.size = 1,
  expfamily = "gaussian", model.fcn = stats::glm,
  PartitioningVariables = "", ...)
}
\arguments{
\item{Data}{data table containing columns of "Species", 'N' (counts, 0<=N<=3), "Sample".}

\item{X}{meta-data containing "Sample" and variables found in \code{frmla}}

\item{frmla}{Formula. If \code{expfamily='binomial'}, must have c(Successes,Failures)~. Otherwise, the variable for data is "Data", e.g. \code{Data~phylo}. Explanatory variables used for phylofactorization must interact with the variable \code{phylo}, e.g. partitioning by pH and not N requires \code{Data~N+phylo*pH}}

\item{nfactors}{integer for number of factors to find}

\item{ncores}{integers for number of cores to use for parallelization}

\item{binom.size}{binom.size of binomial samples for each species. binom.size=1 for presence/absence data.}

\item{expfamily}{Either "gaussian" or "binomial" - determines the aggregation method.}

\item{model.fcn}{Regression function, such as glm, gam, glm.nb, gls. Must have column labelled "Deviance" in \code{\link{anova}}.}

\item{PartitioningVariables}{Character vector containing the variables in \code{frmla} to be used for phylofactorization. Objective function will be the sum of deviance from all variables listed here.}
}
\description{
Generalized phylofactorization - currently skeleton for binomial glm; will expand to exponential, gamma etc. soon.
}
\examples{
library(phylofactor)

set.seed(1)
m <- 50
n <- 200
tree <- rtree(m)
X <- data.table('y'=rnorm(n),
                'z'=rnorm(n,sd=0.5),
                'Sample'=sapply(1:n,FUN=function(s) paste('Sample',s)),
                key='Sample')
#we'll partition by 'y'.
binom.size=3
clade <- phangorn::Descendants(tree,75,'tips')[[1]]
clade2 <- phangorn::Descendants(tree,53,'tips')[[1]]

######## presence/absence dataset with affected clade #######
## most species have higher P{present} with y
ilogit <- function(eta) 1/(1+exp(-eta))
eta <- X$z+X$y
p <- ilogit(eta)
M <- matrix(rbinom(m*n,binom.size,rep(p,times=m)),nrow=m,ncol=n,byrow=T)
rownames(M) <- tree$tip.label
colnames(M) <- X$Sample

#### the first clade decreases with y ####
eta1 <- X$z-X$y
p1 <- ilogit(eta1)
for (species in clade){
   M[species,] <- rbinom(n,binom.size,p1)
}
#### the second clade weakly decreases with y ####
eta2 <- X$z-.3*X$y
p2 <- ilogit(eta2)
for (species in clade2){
   M[species,] <- rbinom(n,binom.size,p2)
}

####################### to partition on y, must have phylo* #########
pf <- gpf(M,tree,X,frmla=cbind(Successes,Failures)~z+phylo*y,nfactors=2,binom.size=binom.size,family=binomial(link='logit'))
all.equal(pf$groups[[1]][[1]],clade)

pf.tree(pf)
par(mfrow=c(2,1))
phytools::phylo.heatmap(tree,M[,order(X$y)])
phytools::phylo.heatmap(tree,ilogit(pf.predict(pf)[,order(X$y)]))

################# Poisson Regression
eta <- .3*X$z
lambda <- exp(eta)
M <- matrix(rpois(m*n,rep(lambda,times=m)),nrow=m,ncol=n,byrow=T)
rownames(M) <- tree$tip.label
colnames(M) <- X$Sample

#### the first clade decreases with y ####
eta1 <- .3*X$z-X$y
lambda1 <- exp(eta1)
for (species in clade){
   M[species,] <- rpois(n,lambda1)
}
#### the second clade strongly increases with y ####
eta2 <- .3*X$z-0.3*X$y
lambda2 <- exp(eta2)
for (species in clade2){
   M[species,] <- rpois(n,lambda2)
}



##For non-binomial, use "Data" as response variable #########
pf <- gpf(M,tree,X,frmla=Data~z+phylo*y,nfactors=2,family=poisson)
list(clade,clade2)
pf$bins

par(mfrow=c(2,1))
phytools::phylo.heatmap(tree,M[,order(X$y)])
phytools::phylo.heatmap(tree,exp(pf.predict(pf)[,order(X$y)]))
}
