% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTaxa.R
\name{listTaxa}
\alias{listTaxa}
\title{Lists all unique taxonomic names in a Taxonomy at a given level}
\usage{
listTaxa(taxonomy, level = "p")
}
\arguments{
\item{taxonomy}{Taxonomy. First row is species IDs, second row is their taxonomy. This function assumes greengenes-compatible taxonomic strings}

\item{level}{Taxonomic level from set {'k','p','c','o','f','g'}}
}
\value{
list of taxonomic strings for all taxa up to speficied level.
}
\description{
Lists all unique taxonomic names in a Taxonomy at a given level
}
\examples{
data(FTmicrobiome)
listTaxa(FTmicrobiome$taxonomy,level='p')
}
