% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mAggregation.R
\name{mAggregation}
\alias{mAggregation}
\title{marginally stable aggregation of binomial data}
\usage{
mAggregation(DF, grp, tree, X, binom.size, expfamily)
}
\arguments{
\item{DF}{data table containing "Species", "sample", and "N" (N= counts of species in sample)}

\item{grp}{see output of \code{\link{getPhyloGroups}}}

\item{tree}{phylo object}

\item{X}{meta-data table containing "sample" and variables found in formula}

\item{binom.size}{Size of binomial (binom.size=1 for Bernoulli)}

\item{expfamily}{Character, currently either "binomial" or "gaussian", indicating family of exponential family for marginally-stable aggregation.}
}
\description{
marginally stable aggregation of binomial data
}
