% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mAggregation.R
\name{mAggregation}
\alias{mAggregation}
\title{marginally stable aggregation of binomial data}
\usage{
mAggregation(DF, grp, tree, X, size = 1, SampleFrame = NULL)
}
\arguments{
\item{DF}{data table containing "Species", "sample", and "N" (N= counts of species in sample)}

\item{grp}{see output of \code{\link{getGroups}}}

\item{tree}{phylo object}

\item{X}{meta-data table containing "sample" and variables found in formula}

\item{SampleFrame}{Data table containing "sample" and "G" indicating all possible combinations of samples and groups.}
}
\description{
marginally stable aggregation of binomial data
}
