% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.TipTest.R
\name{pf.TipTest}
\alias{pf.TipTest}
\title{Test of whether tips are over/under-represented in phylofactorization. Currently in-development.}
\usage{
pf.TipTest(PF, factors = 1:(PF$nfactors), lower.tail = FALSE)
}
\arguments{
\item{PF}{phylofactor object. See \code{\link{PhyloFactor}}}

\item{factors}{Set of factors to test. Default is all the factors in PF input.}

\item{lower.tail}{Logical, whether test should be one-tailed test looking at lower tail.}
}
\description{
Performs hypergeometric test on the number of tips in a phylofactorization
}
