% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.tree.R
\name{pf.tree}
\alias{pf.tree}
\title{ggtree-based plotting of phylofactor bins or factors}
\usage{
pf.tree(pf, method = "factors", factor.map = NULL, Grps = NULL,
  bg.color = NA, bg.alpha = 0.1, alphas = NULL, layout = "circular",
  rootnode = FALSE, color.fcn = rainbow, ...)
}
\arguments{
\item{pf}{phylofactor class object}

\item{method}{either "factors" or "bins", based on whether to use factors (and \code{factor.map} or resultant bins from phylofactorization)}

\item{factor.map}{data frame with two columns. The first column is the factor number and the second is which group (1 or 2) should be highlighted}

\item{Grps}{optional list of groups for plotting}

\item{bg.color}{background color}

\item{bg.alpha}{background alpha}

\item{alphas}{vector, of length = \code{nrow(factor.map)} or \code{length(Grps)} or, if default settings, of length \code{pf$nfactors}. Controls alphas for highlights of each clade}

\item{layout}{See \code{\link{ggtree}}. Default is "circular"}

\item{rootnode}{Logical. If true, will fill in the root node. Default controls root only with bg.color}

\item{color.fcn}{Color function for plotting. Default is rainbow.}

\item{...}{optional arguments for ggtree}
}
\description{
ggtree-based plotting of phylofactor bins or factors
}
\examples{
library(phylofactor)
data(FTmicrobiome)
PF <- FTmicrobiome$PF
gg <- pf.tree(PF,factor.map=data.frame('factors'=1:3,'groups'=rep(1,3)),layout='circular')
gg$ggplot
## legend: PF$groups[gg$map] <--> gg$colors
}
