% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.tree.R
\name{pf.tree}
\alias{pf.tree}
\title{ggtree-based plotting of phylofactor bins or factors}
\usage{
pf.tree(pf, method = "factors", universal.tree = NULL, factors = NULL,
  groups = NULL, colors = NULL, GroupList = NULL, bg.color = NA,
  bg.alpha = 0.1, alphas = NULL, layout = "circular", rootnode = FALSE,
  top.layer = F, top.alpha = 0.1, color.fcn = viridis::viridis, ...)
}
\arguments{
\item{pf}{phylofactor class object}

\item{method}{either "factors" or "bins", based on whether to use factors ( \code{factors} along with \code{groups}) or resultant bins (\code{pf$bins}) from phylofactorization.}

\item{universal.tree}{A tree containing all the tips in \code{pf$tree}. Tips in \code{pf$tree} will be coded as dots}

\item{factors}{vector of factor numbers to be plotted}

\item{groups}{vector of groups (1 or 2) to be plotted for each factor. Default is 1}

\item{colors}{colors for each \code{factors x groups} pair, or each member of \code{GroupList}}

\item{GroupList}{optional list of groups for plotting}

\item{bg.color}{background color}

\item{bg.alpha}{background alpha}

\item{alphas}{vector, of length = \code{nrow(factor.map)} or \code{length(GroupList)} or, if default settings, of length \code{pf$nfactors}. Controls alphas for highlights of each clade. Warning: the current output legend gives you the colors layered onto each clade. Layering transparent colors on top of one-another will produce a different color not in the legend..}

\item{layout}{See \code{\link{ggtree}}. Default is "circular"}

\item{rootnode}{Logical. If true, will fill in the root node. Default controls root only with bg.color}

\item{color.fcn}{Color function for plotting. Default is viridis.}

\item{...}{optional arguments for ggtree}
}
\description{
ggtree-based plotting of phylofactor bins or factors
}
\examples{
library(phylofactor)
data(FTmicrobiome)
pf <- FTmicrobiome$PF
gg <- pf.tree(pf,factors=1:3,layout='circular')
gg$ggplot
gg$legend
}
