% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglm.R
\name{pglm}
\alias{pglm}
\title{Returns regression of x on y according to frmla}
\usage{
pglm(y, x, frmla, smallglm = F, ...)
}
\arguments{
\item{y}{response variable}

\item{x}{independent variable}

\item{frmla}{Formula for dependence of y on x}
}
\value{
glm object
}
\description{
Returns regression of x on y according to frmla
}
\examples{
x <- rnorm(10)
y <- x+rnorm(10)
phyloreg(y,x,y~x)
}

