% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloFactor.R
\name{PhyloFactor}
\alias{PhyloFactor}
\title{Performs phylofactorization of a compositional dataset}
\usage{
PhyloFactor(Data, tree, X, frmla = NULL, method = "ILR", choice = "var",
  Grps = NULL, nclades = NULL, stop.fcn = NULL, stop.early = NULL,
  ncores = NULL, clusterage = 1, ...)
}
\arguments{
\item{Data}{Data matrix whose rows are tip labels of the tree, columns are samples of the same length as X, and whose columns sum to 1}

\item{tree}{Phylogeny whose tip-labels are row-names in Data.}

\item{X}{independent variable.}

\item{frmla}{Formula for input in GLM. Default formula is Data ~ X}

\item{method}{Method for amalgamating groups and constructing basis. Default method = "ILR" uses the isometric log-ratio transform. Coming soon: method='add' which uses additive grouping with log-ratio regression.}

\item{choice}{Choice, or objective, function for determining the best edges at each iteration. Must be choice='var' or choice='F'. 'var' minimizes residual variance, whereas 'F' maximizes the F-statistic from an analysis of variance.}

\item{Grps}{Optional input of groups to be used in analysis to override the groups used in Tree. for correct format of groups, see output of getGroups}

\item{nclades}{Number of clades or factors to produce in phylofactorization. Default, NULL, will iterate phylofactorization until either dim(Data)[1]-1 factors, or until stop.fcn returns T}

\item{stop.fcn}{Currently, accepts input of 'KS'. Coming soon: input your own function of the environment in phylofactor to determine when to stop.}

\item{stop.early}{Logical indicating if stop.fcn should be evaluated before (stop.early=T) or after (stop.early=F) choosing an edge maximizing the objective function.}

\item{ncores}{Number of cores for built-in parallelization of phylofactorization. Parallelizes the extraction of groups, amalgamation of data based on groups, regression, and calculation of objective function. Be warned - this can lead to R taking over a system's memory, so see clusterage for how to control the return of memory from clusters.}

\item{clusterage}{Age, i.e. number of iterations, for which a phyloFcluster should be used before returning its memory to the system. Default age=1.}
}
\value{
Phylofactor object, a list containing: "method", "Data", "tree" - inputs from phylofactorization. Output also includes "factors","glms","terminated" - T if stop.fcn terminated factorization, F otherwise - "atoms", "atom.sizes", "basis" - basis from "method" for projection of data onto phylofactors, and "Monophyletic.Clades" - a list of which atoms are monophyletic and have atom.size>1
}
\description{
Performs phylofactorization of a compositional dataset
}
\examples{
 set.seed(1)
 library(ape)
 library(phangorn)
tree <- unroot(rtree(7))
X <- as.factor(c(rep(0,5),rep(1,5)))
sigClades <- Descendants(tree,c(9,12),type='tips')
Data <- matrix(rlnorm(70,meanlog = 8,sdlog = .5),nrow=7)
rownames(Data) <- tree$tip.label
colnames(Data) <- X
Data[sigClades[[1]],X==0] <- Data[sigClades[[1]],X==0]*8
Data[sigClades[[2]],X==1] <- Data[sigClades[[2]],X==1]*9
Data <- t(clo(t(Data)))
Atoms <- atoms(G=sigClades,set=1:7)
PF <- PhyloFactor(Data,tree,X,nclades=2)
PF$atoms
all(PF$atoms \%in\% Atoms)
}

