% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylofactor.summary.R
\name{phylofactor.summary}
\alias{phylofactor.summary}
\title{Summary of phylofactor object for a given node or factor number}
\usage{
phylofactor.summary(PF, taxonomy, factor = NULL, tree = PF$tree,
  simplify.TaxaIDs = F, subtree = F, prediction = T, tipLabels = F,
  Transform = "atan", ...)
}
\arguments{
\item{PF}{PhyloFactor object}

\item{taxonomy}{Taxonomy, first column is OTU ids in tree, second column is greengenes taxonomic string}

\item{factor}{Factor number to summarize.}

\item{tree}{phylogeny (ape class)}

\item{simplify.TaxaIDs}{Whether or not to simplify IDs of group taxonomy to their shorest unique prefix (for each OTU in each group, this is taxonomy down to the coarsest taxonomic level unique to the OTU's group)}

\item{subtree}{Logical indicating whether or not to output the subtree partitioned when PF$node was split (i.e. sub-tree split by factor)}

\item{prediction}{Logical. If subtree=T, prediction=T will produce a phylo.heatmap containing the predicted data. Otherwise, will output phylo.heatmap of real data}

\item{tipLabels}{Logical indicating whether or not to include tip labels in plot of subtree.}

\item{Transform}{String, method for transforming data for easier visualization of sub-tree phylo.heatmaps. Options include "atan" to flatten extreme values, "tan" to amplify extreme values, and "clr" for centered log-ratio transform.}

\item{...}{additional arguments for phylo.Heatmap. If phyloHeatmap='AW', these are additional arguments for \code{plot.phylo}. Otherwise, see description for \code{phylo.heatmap} in phytools package for list of additional arguments.}

\item{node}{node number, must be in PF$nodes. Takes priority over "factor" for what is summarized}
}
\value{
summary object. List containing $group and $complement info, each containing summary.group output for that group -  $IDs, $otuData and $PF.prediction
}
\description{
Summary of phylofactor object for a given node or factor number
}
\examples{
data("FTmicrobiome")
OTUTable <- FTmicrobiome$OTUTable        #OTU table
Taxonomy <- FTmicrobiome$taxonomy        #taxonomy
tree <- FTmicrobiome$tree                #tree
X <- FTmicrobiome$X                      #independent variable - factor indicating if sample is from feces or tongue

rm('FTmicrobiome')

# remove rare taxa
ix <- which(rowSums(OTUTable==0)<30)
OTUTable <- OTUTable[ix,]
OTUs <- rownames(OTUTable)
tree <- drop.tip(tree,which(!(tree$tip.label \%in\% OTUs)))

par(mfrow=c(1,1))
phylo.heatmap(tree,t(clr(t(OTUTable))))
PF <- PhyloFactor(OTUTable,tree,X,nfactors=2,choice='var')

FactorSummary <- phylofactor.summary(PF,Taxonomy,factor=1)

str(FactorSummary)

par(mfrow=c(1,2))
plot(FactorSummary$ilr,ylab='ILR coordinate',main='ILR coordinate of factor',xlab='sample no.',pch=16)
lines(FactorSummary$fitted.values,lwd=2,col='blue')
legend(x=1,y=-5,list('data','prediction'),pch=c(16,NA),lty=c(NA,1),col=c('black','blue'),lwd=c(NA,2))

plot(FactorSummary$MeanRatio,ylab='ILR coordinate',main='Mean Ratio of Grp1/Grp2',xlab='sample no.',pch=16)
lines(FactorSummary$fittedMeanRatio,lwd=2,col='blue')
legend(x=1,y=-5,list('data','prediction'),pch=c(16,NA),lty=c(NA,1),col=c('black','blue'),lwd=c(NA,2))
}

