% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloRegression.R
\name{PhyloRegression}
\alias{PhyloRegression}
\title{Performs regression of groups of Data against X for an input list of groups, and outputs the best group based on "choice" function input.}
\usage{
PhyloRegression(LogData, X, frmla, Grps, choice, treeList, cl,
  totalvar = NULL, ix_cl, treetips, grpsizes, tree_map, quiet = T,
  nms = NULL, ...)
}
\arguments{
\item{LogData}{Logarithm of Compositional data matrix whose rows are parts and columns are samples.}

\item{X}{independent variables for input into glm.}

\item{frmla}{Formula for input into glm by lapply(Y,FUN = pglm,x=X,frmla=frmla,choice,...).}

\item{Grps}{Groups - a list whose elements are two-element lists containing the groups and their compliments for amalgamation into ILR coordinates}

\item{choice}{Choice function for determining the group maximizing the objective function. Currently the only allowable inputs are 'var' - minimize residual varaince - and 'F' - minimize test-statistic from anova.}

\item{cl}{phyloFcluster input for built-in parallelization of grouping, amalgamation, regression, and objective-function calculation.}
}
\description{
Performs regression of groups of Data against X for an input list of groups, and outputs the best group based on "choice" function input.
}
\examples{
data("FTmicrobiome")
library(ape)
library(compositions)
Y <- FTmicrobiome$OTUTable
Y <- Y[which(rowSums(Y==0)<30),] #only include taxa present in at least 29 samples
Y[Y==0]=.65

Y <- t(clo(t(Y)))
tree <- drop.tip(FTmicrobiome$tree, setdiff(FTmicrobiome$tree$tip.label,rownames(Y)))
X <- FTmicrobiome$X
Grps <- getGroups(tree)
choice='var'
cl <- phyloFcluster(2)

frmla <- Y~X
print(head(Y))
Z=Y

pr <- PhyloRegression(Z,X,frmla,Grps,choice,cl)

stopCluster(cl)
gc()
par(mfrow=c(1,2))
image(clr(t(Y)),main="Original Data")
image(clr(t(pr$residualData)),main="residual Data")
}

