<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.group.R
\name{summary.group}
\alias{summary.group}
\title{Summarize groups from phylofactor}
\usage{
\method{summary}{group}(PF, tree, taxonomy, factor, grp)
}
\arguments{
\item{PF}{Phylofactor output}

\item{tree}{The tree}

\item{taxonomy}{The taxonomy.}

\item{factor}{The factor.}

\item{grp}{The group.}
}
\description{
This is the description
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.group.R
\name{summary.group}
\alias{summary.group}
\title{Summarizes OTU IDs, taxonomic detail, data and phylofactor predictions of a given input group of taxa}
\usage{
summary.group(PF, tree, taxonomy, factor, grp)
}
\arguments{
\item{PF}{PhyloFactor object}

\item{tree}{phylogeny (ape class)}

\item{taxonomy}{First column is OTU ids also found in tree$tip.label and second column is greengenes taxonomic strings}

\item{factor}{Factor number up to which phylofactor predictions should be used}

\item{grp}{input vector of taxa to be summarized}
}
\value{
summary object with $IDs, $otuData, $PF.prediction
}
\description{
Summarizes OTU IDs, taxonomic detail, data and phylofactor predictions of a given input group of taxa
>>>>>>> ba8c56008d4af4096484046051c47926207fc9ec
}

