% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.group.R
\name{summary.group}
\alias{summary.group}
\title{Summarizes OTU IDs, taxonomic detail, data and phylofactor predictions of a given input group of taxa}
\usage{
\method{summary}{group}(PF, tree, taxonomy, factor, grp, simplify = F)
}
\arguments{
\item{PF}{PhyloFactor object}

\item{tree}{phylogeny (ape class)}

\item{taxonomy}{First column is OTU ids also found in tree$tip.label and second column is greengenes taxonomic strings}

\item{factor}{Factor number up to which phylofactor predictions should be used}

\item{grp}{input vector of taxa to be summarized}
}
\value{
summary object with $IDs, $otuData, $PF.prediction
}
\description{
Summarizes OTU IDs, taxonomic detail, data and phylofactor predictions of a given input group of taxa
}

