% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylofactor.R
\name{summary.phylofactor}
\alias{summary.phylofactor}
\title{Summarize phylofactor objects}
\usage{
summary.phylofactor(PF, taxonomy = NULL, factor = NULL,
  taxon.trimming = "sup", output.signal = T)
}
\arguments{
\item{PF}{phylofactor object}

\item{taxonomy}{taxonomy (optional) for taxonomic trimming. First column contains tree tip-labels, second column has semicolon-delimited taxonomic character strings.}

\item{factor}{Integer (optional). If NULL, will use print.phylofactor}

\item{taxon.trimming}{Chracter string. How to trim taxonomic groups. Must be either "sup" (group by shortest unique prefix for every taxon in group distinguishing it from taxa in complementary group), "taxon" for taxon levels in second column of \code{taxonomy}, or "species" for no taxonomic binning.}

\item{output.signal}{logical - whether or not to output signal. Will attempt to call model.fcn or customized choice.fcn to estimate signal (and thus relative importance) for each taxonomic group from \code{taxon.trimming}.}
}
\value{
phylofactor.summary object containing data, group summary and, with input taxonomy, taxonomic break-down of signal
}
\description{
Summarize phylofactor objects
}
