% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phylofactor.r
\name{summary.phylofactor}
\alias{summary.phylofactor}
\title{Summary of phylofactor object for a given node or factor number}
\usage{
\method{summary}{phylofactor}(PF, tree, taxonomy, node = NULL,
  factor = NULL, subtree = F, prediction = T, tipLabels = F, ...)
}
\arguments{
\item{PF}{PhyloFactor object}

\item{tree}{phylogeny (ape class)}

\item{taxonomy}{Taxonomy, first column is OTU ids in tree, second column is greengenes taxonomic string}

\item{node}{node number, must be in PF$nodes. Takes priority over "factor" for what is summarized}

\item{factor}{Factor number to summarize.}

\item{subtree}{Logical indicating whether or not to output the subtree partitioned when PF$node was split (i.e. sub-tree split by factor)}

\item{prediction}{Logical. If subtree=T, prediction=T will produce a phylo.heatmap containing the predicted data. Otherwise, will output phylo.heatmap of real data}

\item{tipLabels}{Logical indicating whether or not to include tip labels in plot of subtree.}
}
\value{
summary object. List containing $group and $complement info, each containing summary.group output for that group -  $IDs, $otuData and $PF.prediction
}
\description{
Summary of phylofactor object for a given node or factor number
}
\examples{
data("FTmicrobiome")
OTUTable <- FTmicrobiome$OTUTable        #OTU table
Taxonomy <- FTmicrobiome$taxonomy        #taxonomy
tree <- FTmicrobiome$tree                #tree
X <- FTmicrobiome$X                      #independent variable - factor indicating if sample is from feces or tongue

rm('FTmicrobiome')

# remove rare taxa
ix <- which(rowSums(OTUTable==0)<30)
OTUTable <- OTUTable[ix,]
OTUs <- rownames(OTUTable)
tree <- drop.tip(tree,which(!(tree$tip.label \%in\% OTUs)))
delta=0.65
OTUTable[OTUTable==0]=delta
OTUTable <- OTUTable \%>\% t \%>\% clo \%>\% t
OTUTable <- OTUTable[tree$tip.label,]

par(mfrow=c(1,1))
phylo.heatmap(tree,t(clr(t(OTUTable))))
PF <- PhyloFactor(OTUTable,tree,X,nclades=2,choice='var')

FactorSummary <- summary.phylofactor(PF,tree,Taxonomy,factor=1)
NodeSummary <- summary.phylofactor(PF,tree,Taxonomy, node=PF$nodes[2])

str(FactorSummary)

NodeSummary$group$IDs

}

