% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaTree.R
\name{TaxaTree}
\alias{TaxaTree}
\title{Construct tree for particular taxonomic group}
\usage{
TaxaTree(tree, taxonomy, Taxon)
}
\arguments{
\item{tree}{Ape class phylogeny whose tips are contained in the OTUids of the taxonomy}

\item{taxonomy}{Taxonomy list whose first column is OTUids and whose second column is taxonomic strings}

\item{Taxon}{the particular Taxonomic string to include in tree.}
}
\value{
tree whose tips are all the elements in the tree with taxonomy containing the string \code{Taxon}
}
\description{
Construct tree for particular taxonomic group
}

