% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaTree.R
\name{taxaTree}
\alias{taxaTree}
\title{make tree from vector of standardized taxonomies}
\usage{
taxaTree(taxa)
}
\arguments{
\item{taxa}{character vector of semicolon-delimited taxonomic names}
}
\description{
make tree from vector of standardized taxonomies
}
\examples{
taxa <- c('Prokarya; Bacteria; Proteobacteria; rando1',
          'Prokarya; Bacteria; Proteobacteria; rando2',
          'Eukarya; Animalia; Vertebrata; cat',
          'Eukarya; Animalia; Vertebrata; dog',
          'Eukarya; Fungi; Ascomycota; yeast')
tree <- taxaTree(taxa)
plot(tree)
}
