% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueTaxa.R
\name{uniqueTaxa}
\alias{uniqueTaxa}
\title{Find shortest unique prefix in two taxonomic lists x and y}
\usage{
uniqueTaxa(x, y)
}
\arguments{
\item{x}{list of taxonomic strings whose unique taxonomies we want to know}

\item{y}{list of taxonomic strings for "subtraction" from x}
}
\value{
list of taxonomic strings unique to x, i.e. the shortest taxonomic prefix of x that is unique to x
}
\description{
Find shortest unique prefix in two taxonomic lists x and y
}

